requirements-dev.txt
fastapi==0.121.2
uvicorn==0.38.0
prometheus_client==0.23.1
httpx==0.28.1
SQLAlchemy==2.0.44
asyncpg==0.30.0
greenlet==3.2.4
pydantic==2.12.4
structlog==25.5.0
pytest-asyncio==1.3.0
pytest-mock==3.15.1
tenacity==9.0.0